+++ src/main/java/com/sochoeun/exception/NotFoundException.java
@@ -3,7 +3,7 @@ package com.sochoeun.exception;
 import org.springframework.http.HttpStatus;
 
 public class NotFoundException extends ApiException{
-    public NotFoundException(String resourceName,Integer id) {
+    public NotFoundException(String resourceName,Long id) {
         super(HttpStatus.NOT_FOUND,String.format("%s with ID : %d Not Found",resourceName,id));
     }
 }
