+++ src/main/java/com/sochoeun/repository/ModelRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ModelRepository extends JpaRepository<Model,Integer> {
+public interface ModelRepository extends JpaRepository<Model,Long> {
     Page<Model> findAllByBrandId(Integer brandId, Pageable pageable);
 }
